HIPP-618 refact: move config files
